Fix container start sequence
Warning!  Major hack alert.

When we create a crio-lxc container, we start the lxc container to
wait for the syncfifo.  The crio-lxc container isn't really started
until the fifo is written to.

So we can't just use c.Running() to tell us that the crio-lxc container
is running.

Ideally we could add a 'criolxc.state' key in the container config.
Lxc would happily ignore that.  However, we cannot use the lxc api
to do that because it doesn't know about the keys.  And if we do it
ourselves, then we need to worry about locking.

So, add a pre-start hook of /bin/true to the config file when the
syncfifo is unlocked.  If anyone wants to set /bin/true as a real
pre-start hook, come talk to me.

Other ideas are definitely welcome!

Signed-off-by: Serge Hallyn <shallyn@cisco.com>